Flutter no logger

+++ lib/models/inventory_item_v2.dart
@@ -101,7 +101,7 @@ class Id {
 @JsonSerializable()
 class Terms {
   @JsonKey(name: "priceEach")
-  int? priceEach;
+  double? priceEach;
   @JsonKey(name: "pvEach")
   int? pvEach;
